requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
click==8.1.2
dnspython==2.2.1
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.28
protobuf==3.20.0
pymongo==4.1.1
soupsieve==2.3.2.post1
Werkzeug==2.1.1
zipp==3.8.0